把複製函式放到utils/。新增spec。
