+++ src/元素/翻譯/翻譯結果.jsx
@@ -7,27 +7,7 @@ import {
   意傳服務,
 } from "demo-ui";
 import Container漢羅列表 from "../顯示/漢羅列表.container";
-
-export const 計算複製內容 = (綜合標音 = []) => {
-  if (!綜合標音 || 綜合標音.length < 1) {
-    return 綜合標音;
-  }
-
-  return 綜合標音
-  .map((item) => {
-    const 漢字 = item.漢字.replace(/ /g, "");
-    return {
-      漢字臺羅: [漢字, item.臺灣客話].join("\n"),
-      臺羅: item.臺灣客話,
-      漢字,
-    };
-  })
-  .reduce((acc, item) => ({
-    漢字臺羅: [acc.漢字臺羅, item.漢字臺羅].join("\n"),
-    漢字: [acc.漢字, item.漢字].join("\n"),
-    臺羅: [acc.臺羅, item.臺羅].join("\n"),
-  }));
-};
+import 計算複製內容 from "../../utils/複製";
 
 class 翻譯結果 extends React.Component {
   取得複製鈕群() {
